upgrade image icons
