+++ native_device_feature/lib/screens/place_details_screen.dart
@@ -1,23 +1,1 @@
-import 'package:flutter/material.dart';
-import 'package:native_device_feature/screens/add_place_screen.dart';
 
-class PlacesListScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Your places'),
-        actions: [
-          IconButton(
-              icon: Icon(Icons.add),
-              onPressed: () {
-                Navigator.of(context).pushNamed(AddPlaceScreen.routeName);
-              })
-        ],
-      ),
-      body: Center(
-        child: CircularProgressIndicator(),
-      ),
-    );
-  }
-}
