+++ native_device_feature/lib/screens/places_list_screen.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+import '../screens/add_place_screen.dart';
+
+class PlacesListScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Your places'),
+        actions: [
+          IconButton(
+              icon: Icon(Icons.add),
+              onPressed: () {
+                Navigator.of(context).pushNamed(AddPlaceScreen.routeName);
+              })
+        ],
+      ),
+      body: Center(
+        child: CircularProgressIndicator(),
+      ),
+    );
+  }
+}
